WIP: Very basic, likely naive use of module-lexer
This uses a #lang supplied 'color-lexer to set Emacs character syntax
and faces.

Currently this works by disabling font-lock-mode. It might turn out to
be possible to make this work reliably using font-lock, and even to
work faster that way, thanks to Emacs' default jit-lock-mode. However
in my early attempts I was getting many confusing results due to the
interaction among after-change-functions, syntax-propertize, and
jit-lock-mode. So for now I'm minimizing the number of moving parts.

[This does not attempt to use a #lang supplied drracket:indentation.
Indeed I'm skeptical that will turn out to be satisfactory. I suspect
that a new interface will be necessary for that sort of thing.]