Smarter bounding of re-lexing
Re the start, skip ahead in the port to the desired location (working
in terms of port-next-location offsets not raw file position). Use
another interval-map to remember the lexer "mode" in effect from the
previous lex, so we can use it again. (We don't store this in the
`token` struct because that would complicate equality checking.)

Re the stop, have set-interval return a boolean about whether to
continue. That way, we can just stop, as opposed to continuing to lex
all the way to the end while ignoring the results.