Explicitly set process coding to utf-8; see issue #484
Although I'm not yet sure this will fix the issue for the person
reporting it, and therefore not marking this as closing the issue, I
think it might: The only way I have been able to reproduce a similar
problem, so far, is to set the coding wrong intentionally. Possibly
the default process coding is not utf-8 on their system.